fix(config): Adding modularized config settings
