+++ src/lib/network.js
@@ -1,6 +1,6 @@
 const errorMessages = require('./errorMessages')
 const BCHJS = require('@chris.troutner/bch-js')
-const config = require('../../config.json')
+const config = require('../../config')
 
 const bchjs = new BCHJS({
   restURL:
